requirements_raspi.txt
paho-mqtt>=2.1.0
opencv-python>=4.5.5.64